requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
confluent-kafka==2.7.0
cryptography==44.0.0
dotenv==0.9.9
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
flask-redis==0.4.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
scikit-learn==1.6.0
scipy==1.14.1
SQLAlchemy==2.0.36
threadpoolctl==3.5.0
typing_extensions==4.12.2
Werkzeug==3.1.3
